+++ resources/views/livewire/dashboard/top.blade.php
@@ -38,6 +38,9 @@
                     @if ($l->isMysqlService() )
                       {{ $l->configuration->mysqlVersion }}
                     @endif
+                    @if ($l->isPostgresqlService() )
+                      {{ $l->configuration->postgresqlVersion }}
+                    @endif
 
                   </span>