+++ resources/views/yaml/set_env.blade.php
@@ -1,6 +1,6 @@
 # Set environment
       env:
-@if ( $databaseType === "mysql" )
+@if ( $databaseType === "mysql")
 @if ( $mysqlDatabase != "")
         DB_CONNECTION: {{ $mysqlDatabase }}
 @endif
@@ -18,6 +18,21 @@
 @endif
 @if ( $mysqlPasswordType === 'hardcoded' )
         DB_PASSWORD: {{ $mysqlPassword }}
+@endif
+@elseif ( $databaseType === "postgresql")
+        DB_CONNECTION: pgsql
+@if ( $postgresqlDatabaseName != "")
+        DB_DATABASE: {{ $postgresqlDatabaseName }}
+@endif
+@if ( $postgresqlDatabasePort != "")
+        DB_PORT: {{ $postgresqlDatabasePort }}
+@endif
+        DB_USERNAME: postgres
+@if ( $postgresqlPasswordType === 'secret' )
+        DB_PASSWORD: $@{{ secrets.DB_PASSWORD }}
+@endif
+@if ( $postgresqlPasswordType === 'hardcoded' )
+        DB_PASSWORD: {{ $postgresqlPassword }}
 @endif
 @else
         SESSION_DRIVER: array